feat: Verify sshtunnel does not exists

+++ mason.sh
@@ -249,6 +249,23 @@ else
     exit 1
 fi
 
+# ----------------------------- sshtunnel.service ---------------------------- #
+
+# Check if sshtunnel service exsists, if not create it.
+service_exists() {
+    local n=$1
+    if [[ $(systemctl list-units --all -t service --full --no-legend "$n.service" | sed 's/^\s*//g' | cut -f1 -d' ') == $n.service ]]; then
+        return 0
+    else
+        return 1
+    fi
+}
+
+if service_exists sshtunnel; then
+    echo "sshtunnel.service exsists, exiting mason"
+    exit 0
+fi
+
 cat <<EOF > /etc/systemd/system/sshtunnel.service
 [Unit]
 Description=Service to maintain an ssh reverse tunnel
